api/requirements/base.txt
fastapi==0.115.11
pydantic-settings==2.8.1
uvicorn==0.34.0
PyJWT==2.10.1
